requirements.txt
fastapi==0.135.1
fitz==0.0.1.dev2
langchain_community==0.4.1
langchain_core==1.2.19
langchain_huggingface==1.2.1
langchain_text_splitters==1.1.1
matplotlib==3.10.8
numpy==2.4.3
pandas==3.0.1
pydantic==2.12.5
python-dotenv==1.2.2
rapidocr_onnxruntime==1.2.3
Requests==2.32.5
seaborn==0.13.2
sentence_transformers==5.2.3
streamlit==1.55.0
zhipuai==2.1.5.20250825
